chore: add session management, date input, UI and security fixes

+++ controllers/expenseController.js
@@ -2,7 +2,10 @@ const Expense = require("../models/expenseModel");
 
 exports.addExpense = async (req, res) => {
     try {
-        const expense = new Expense(req.body);
+        const expense = new Expense({
+            ...req.body,
+            userId: req.user.id
+        });
         await expense.save();
         res.status(201).json(expense);
     }
@@ -13,7 +16,9 @@ exports.addExpense = async (req, res) => {
 
 exports.getExpenses = async (req, res) => {
     try {
-        const expenses = await Expense.find();
+        const expenses = await Expense.find({
+            userId: req.user.id
+        });
         res.status(200).json(expenses);
     }
     catch (err) {
@@ -23,7 +28,15 @@ exports.getExpenses = async (req, res) => {
 
 exports.deleteExpense = async (req, res) => {
     try {
-        await Expense.findByIdAndDelete(req.params.id);
+        const deleted = await Expense.findOneAndDelete({
+            _id: req.params.id,
+            userId: req.user.id
+        });
+
+        if (!deleted) {
+            return res.status(404).json({ message: "Expense not found" });
+        }
+
         res.status(200).json({ message: "Expense deleted" });
     }
     catch (err) {
@@ -33,10 +46,9 @@ exports.deleteExpense = async (req, res) => {
 
 exports.deleteAllExpenses = async (req, res) => {
     try {
-        await Expense.deleteMany({});
+        await Expense.deleteMany({ userId: req.user.id });
         res.status(200).json({ message: "All expenses deleted" });
-    }
-    catch (err) {
+    } catch (err) {
         res.status(500).json({ message: err.message });
     }
 };
